+++ crates/cubecl-runtime/Cargo.toml
@@ -91,7 +91,6 @@ tracy-client = { workspace = true, optional = true }
 wasm-bindgen-futures = { workspace = true }
 
 [dev-dependencies]
-cubecl-core = { path = "../cubecl-core", version = "=0.11.0-pre.3" }
 rand = { workspace = true, features = ["thread_rng"] }
 serde_json = { workspace = true, features = ["std"] }
 serial_test = { workspace = true }
